+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to
   was too small to give reasonable benchmarking results.
 - Renamed the `europarl2` dataset to `europarl_sent`
 
-## Removed
+### Removed
 - Removed the `europarl1` dataset, as it was too small to give reliable
   benchmarking results. This dataset could not simply be added to the
   `europarl2` dataset, as with the new `lcc` dataset, as the annotaters are not